chore: de-public cells
